added motion to gamelib
