+++ app/gamelib/Cargo.toml
@@ -11,7 +11,8 @@ ogc-rs = {git = "https://github.com/rust-wii/ogc-rs.git", branch = "develop", op
 hecs = { version = "0.9.0", default-features = false }
 rand = { version = "0.8.4", features = ["small_rng"], default-features = false }
 num = { version = "0.4.0", default-features = false}
-
+hashbrown = "0.12"
+micromath = { version = "2.0.0"} # mathematical functions for f32
 
 [features]
 default = []
