+++ app/gamelib/src/lib.rs
@@ -11,3 +11,4 @@ pub use std::{print, println};
 pub mod game_state;
 pub mod game_states;
 pub mod game;
+pub mod plot;
